FAC-340: Create and use SpaceUserAccess domain model

+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/spaceuseraccess/AcceptSpaceInvitationsService.java
@@ -2,6 +2,7 @@ package org.flickit.assessment.users.application.service.spaceuseraccess;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.users.application.domain.SpaceInvitation;
+import org.flickit.assessment.users.application.domain.SpaceUserAccess;
 import org.flickit.assessment.users.application.port.in.spaceinvitee.LoadSpaceUserInvitationsPort;
 import org.flickit.assessment.users.application.port.in.spaceuseraccess.AcceptSpaceInvitationsUseCase;
 import org.flickit.assessment.users.application.port.out.spaceinvitee.DeleteSpaceUserInvitationsPort;
@@ -30,18 +31,19 @@ public class AcceptSpaceInvitationsService implements AcceptSpaceInvitationsUseC
 
         var invitations = loadSpaceUserInvitationsPort.loadInvitations(email);
 
-        List<CreateSpaceUserAccessPort.Param> validInvitations = invitations.stream()
+        List<SpaceUserAccess> validInvitations = invitations.stream()
             .filter(SpaceInvitation::isNotExpired)
-            .map(i -> toUserAccessPortParam(i, param.getUserId())).toList();
-
-        deleteSpaceUserInvitationsPort.deleteAll(email);
+            .map(i -> toSpaceUserAccess(i, param.getUserId())).toList();
 
         if (!validInvitations.isEmpty())
             createSpaceUserAccessPort.persistAll(validInvitations);
+
+        if (!invitations.isEmpty())
+            deleteSpaceUserInvitationsPort.deleteAll(email);
     }
 
-    private CreateSpaceUserAccessPort.Param toUserAccessPortParam(SpaceInvitation invitation, UUID userId) {
-        return new CreateSpaceUserAccessPort.Param(
+    private SpaceUserAccess toSpaceUserAccess(SpaceInvitation invitation, UUID userId) {
+        return new SpaceUserAccess(
             invitation.getSpaceId(),
             userId,
             invitation.getInviterId(),
